fix: prevent kill -TERM -1 in mutation testing on Linux
The stop_daemon_returns_not_running_when_stale_pid test used u32::MAX as
a stale PID. Under the `is_valid_pid -> true` mutant on Linux, this
bypassed validation and caused kill to parse 4294967295 as pid_t -1,
sending SIGTERM to all processes and crashing the CI runner.

Use i32::MAX instead — a valid but non-existent PID that fails harmlessly
with ESRCH. The mutant is still caught by 6 other tests.

+++ src/daemon/pid.rs
@@ -304,7 +304,10 @@ mod tests {
     fn stop_daemon_returns_not_running_when_stale_pid() {
         let dir = tempfile::tempdir().unwrap();
         let path = dir.path().join("stale.pid");
-        write_pid(&path, u32::MAX).unwrap();
+        // Use a valid but non-existent PID (not u32::MAX, which wraps to -1
+        // on Linux and causes `kill -TERM -1` to SIGTERM all processes when
+        // the is_valid_pid -> true mutant is applied).
+        write_pid(&path, i32::MAX as u32).unwrap();
         let result = stop_daemon(&path);
         assert!(matches!(result, Err(DaemonError::NotRunning)));
     }
